+++ test/idGenerator.js
@@ -6,33 +6,46 @@ import {IDGenerator} from "../src/util.js";
 
 describe("IDGenerator", function () {
 
-    describe("#next", function() {
-        const g = new IDGenerator("p");
+    beforeEach(function () {
+        this.gen = new IDGenerator("p");
+    });
 
+    describe("#next", function() {
         it("should prefix IDs", function () {
-            const id = g.next();
+            const id = this.gen.next();
             expect(id).to.startWith("p");
         });
 
         it("should return unique IDs", function () {
-            const id0 = g.next(),
-                  id1 = g.next();
+            const id0 = this.gen.next(),
+                  id1 = this.gen.next();
 
             expect(id0).to.not.equal(id1);
         });
     });
 
-    describe("#increaseToAtLeast", function() {
-        const g = new IDGenerator("p");
+    describe("#avoid", function() {
+        it("should prevent the generator from generating the given ID", function () {
+            this.gen.avoid("p0");
+            const id = this.gen.next();
+            expect(id).to.equal("p1");
+        });
+    });
 
+
+    describe("#increaseToAtLeast", function() {
         it("should set to the maximum of the current value and the given one", function () {
-            g.increaseToAtLeast(10);
-            const id0 = g.next();
+            this.gen.increaseToAtLeast(10);
+            const id0 = this.gen.next();
             expect(id0).to.equal("p10");
 
-            g.increaseToAtLeast(5);
-            const id1 = g.next();
+            this.gen.increaseToAtLeast(5);
+            const id1 = this.gen.next();
             expect(id1).to.equal("p11");
         });
+    });
+
+    after(function () {
+        delete this.gen;
     });
 });